[NAGIOS] Add a plugin to check if we have buildbot master in weird state, with likely infinite build status running (for instance, the trigger).
Addition from Christophe Haen <christophe.haen at cern.ch>: This plugin CANNOT be used as it. You should use it with an event handler that disables
the check when it goes to critical state (so that you keep that state). Then, when you fix the buildmaster issue, you can reenable the plugin and it goes back
to OK. This is needed, because the plugin cannot check whether the found log line really leads to an issue, and cannot know when the issue is fixed.

svn path=/trunk/nagios/; revision=2050